requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.13.0=pypi_0
adabound=0.0.5=pypi_0
addict=2.4.0=pypi_0
aiohttp=3.8.3=pypi_0
aiosignal=1.3.1=pypi_0
albumentations=1.2.1=pypi_0
antlr4-python3-runtime=4.8=pypi_0
argcomplete=1.12.3=pyhd3eb1b0_0
argon2-cffi=20.1.0=py37h27cfd23_1
astunparse=1.6.3=pypi_0
async-timeout=4.0.2=pypi_0
asynctest=0.13.0=pypi_0
attrs=21.4.0=pyhd3eb1b0_0
av=8.1.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.10.0=pypi_0
bias-act-cuda=0.0.0=pypi_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blessings=1.7=pypi_0
boltons=21.0.0=pypi_0
boto3=1.18.40=pypi_0
botocore=1.21.40=pypi_0
bunch=1.0.1=pypi_0
ca-certificates=2022.4.26=h06a4308_0
cached-property=1.5.2=pypi_0
cachetools=4.2.2=pypi_0
certifi=2022.6.15=py37h06a4308_0
cffi=1.15.0=py37hd667e15_1
channelnorm-cuda=0.0.0=pypi_0
charset-normalizer=2.0.4=pypi_0
chumpy=0.70=pypi_0
clang=5.0=pypi_0
click=7.1.2=pypi_0
compress-pickle=2.1.0=pypi_0
correlation-cuda=0.0.0=pypi_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.10.0=pypi_0
cython=0.29.26=pypi_0
debugpy=1.5.1=py37h295c915_0
decorator=4.4.2=pypi_0
defusedxml=0.7.1=pyhd3eb1b0_0
dlib=19.22.1=pypi_0
docker-pycreds=0.4.0=pypi_0
dominate=2.6.0=pypi_0
easydict=1.9=pypi_0
einops=0.4.1=pypi_0
entrypoints=0.3=py37_0
face-alignment=1.3.4=pypi_0
facenet-pytorch=2.5.2=pypi_0
fbpca=1.0=pypi_0
filelock=3.6.0=pypi_0
flatbuffers=1.12=pypi_0
freetype=2.10.4=h5ab3b9f_0
freetype-py=2.2.0=pypi_0
frozenlist=1.3.3=pypi_0
fsspec=2022.11.0=pypi_0
future=0.18.2=pypi_0
fvcore=0.1.5.post20210730=pypi_0
gast=0.4.0=pypi_0
gdown=4.4.0=pypi_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
google-auth=1.34.0=pypi_0
google-auth-oauthlib=0.4.5=pypi_0
google-pasta=0.2.0=pypi_0
gpustat=0.6.0=pypi_0
grpcio=1.39.0=pypi_0
h5py=3.1.0=pypi_0
hessian-eigenthings=0.0.2=pypi_0
hydra-core=1.0.6=pypi_0
idna=3.2=pypi_0
imageio=2.9.0=pypi_0
imageio-ffmpeg=0.4.7=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=4.6.3=pypi_0
importlib-resources=5.7.1=pypi_0
importlib_metadata=4.8.1=hd3eb1b0_0
importlib_resources=5.2.0=pyhd3eb1b0_1
insightface=0.6.2=pypi_0
intel-openmp=2021.3.0=h06a4308_3350
iopath=0.1.9=pypi_0
ipykernel=6.4.1=py37h06a4308_1
ipython=7.29.0=py37hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.4.2=py37_0
jedi=0.18.0=py37h06a4308_1
jinja2=3.0.3=pyhd3eb1b0_0
jmespath=0.10.0=pypi_0
joblib=1.0.1=pypi_0
jpeg=9b=h024ee3a_2
jsonpickle=2.2.0=pypi_0
jsonschema=4.4.0=py37h06a4308_0
jupyter_client=7.0.6=pyhd3eb1b0_0
jupyter_core=4.9.1=py37h06a4308_0
keras=2.6.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=pypi_0
kornia=0.5.10=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
llvmlite=0.36.0=pypi_0
lmdb=1.3.0=pypi_0
lpips=0.1.3=pypi_0
lz4-c=1.9.3=h2531618_0
markdown=3.3.4=pypi_0
markupsafe=2.1.1=py37h7f8727e_0
matplotlib=3.4.2=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py37h14c3975_1001
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.0=py37h42c9631_2
mkl_random=1.2.2=py37h51133e4_0
mmcv=1.6.1=pypi_0
moviepy=1.0.3=pypi_0
multidict=6.0.2=pypi_0
munch=2.5.0=pypi_0
nbconvert=5.5.0=py_0
nbformat=5.3.0=py37h06a4308_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.6.2=pypi_0
ninja=1.10.2=hff7bd54_1
notebook=6.4.11=py37h06a4308_0
numba=0.53.1=pypi_0
numpy=1.19.5=pypi_0
nvdiffrast=0.2.5=pypi_0
nvidia-ml-py3=7.352.0=pypi_0
oauthlib=3.1.1=pypi_0
olefile=0.46=py37_0
omegaconf=2.0.6=pypi_0
onnx=1.12.0=pypi_0
opencv-python=4.6.0.66=pypi_0
opencv-python-headless=4.6.0.66=pypi_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1n=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=21.2=pypi_0
pandas=1.3.1=pypi_0
pandoc=2.12=h06a4308_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.2=pyhd3eb1b0_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.1=py37h2c7a002_0
pip=21.1.3=py37h06a4308_0
portalocker=2.3.0=pypi_0
prettytable=3.3.0=pypi_0
proglog=0.1.10=pypi_0
progressbar=2.5=pypi_0
progressbar2=4.0.0=pypi_0
prometheus_client=0.13.1=pyhd3eb1b0_0
promise=2.3=pypi_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.17.3=pypi_0
psutil=5.9.1=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
py-midi=2.0.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydeprecate=0.3.1=pypi_0
pyglet=1.5.21=pypi_0
pygments=2.10.0=pyhd3eb1b0_0
pyopengl=3.1.0=pypi_0
pyparsing=2.4.7=pypi_0
pyrender=0.1.45=pypi_0
pyrsistent=0.18.0=py37heee7806_0
pyserial=3.5=pypi_0
pysocks=1.7.1=pypi_0
python=3.7.11=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
python-utils=3.3.3=pypi_0
python3-midi=0.2.5=pypi_0
pytorch=1.7.1=py3.7_cuda11.0.221_cudnn8.0.5_0
pytorch-fid=0.2.1=pypi_0
pytorch-lightning=1.5.10=pypi_0
pytz=2021.1=pypi_0
pyvis=0.3.1=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=22.3.0=py37h295c915_2
qudida=0.0.4=pypi_0
readline=8.1=h27cfd23_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
resample2d-cuda=0.0.0=pypi_0
rsa=4.7.2=pypi_0
s3transfer=0.5.0=pypi_0
scikit-image=0.18.2=pypi_0
scikit-learn=0.22=pypi_0
scikit-video=1.1.11=pypi_0
scipy=1.7.0=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
sentry-sdk=1.11.1=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=59.5.0=pypi_0
shapely=1.8.5.post1=pypi_0
shortuuid=1.0.11=pypi_0
six=1.15.0=pypi_0
smmap=5.0.0=pypi_0
soupsieve=2.3.1=pypi_0
sqlite=3.36.0=hc218d9a_0
tabulate=0.8.9=pypi_0
tensorboard=2.6.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorboardx=2.5.1=pypi_0
tensorflow=2.6.2=pypi_0
tensorflow-estimator=2.6.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.13.1=py37h06a4308_0
testpath=0.6.0=py37h06a4308_0
threadpoolctl=2.2.0=pypi_0
tifffile=2021.7.30=pypi_0
tk=8.6.10=hbc83047_0
torch-tb-profiler=0.2.1=pypi_0
torchaudio=0.7.2=py37
torchdiffeq=0.0.1=pypi_0
torchfile=0.1.0=pypi_0
torchmetrics=0.10.3=pypi_0
torchsde=0.2.5=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.8.2=py37_cu110
tornado=6.1=py37h27cfd23_0
tqdm=4.62.0=pypi_0
traitlets=4.3.3=py37h06a4308_0
trampoline=0.1.2=pypi_0
trimesh=3.9.35=pypi_0
typing-extensions=3.7.4.3=pypi_0
upfirdn2d-cuda=0.0.0=pypi_0
urllib3=1.26.12=pypi_0
wandb=0.13.5=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py37_1
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.4.2=py37_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pypi_0
yapf=0.32.0=pypi_0
yarl=1.8.1=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.5.0=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
